adapted parallel algorithm to long int type

+++ matrixMultiplicationP.c
@@ -122,8 +122,8 @@ int main(int argc, char* argv[]) {
             fscanf(inFile, " %d", &matrixB[i][j]);   
     }
 
-    int *sendBuf = multiplyMatricesretVect(partMatrixA, matrixB, rows, colsA, rowsB, colsB, offset, elements);   /*Buffer used to store computed values to send to MASTER*/ 
-    int *recvGatBuf = malloc(sizeof(int) * elements * tasksNum);
+    long int *sendBuf = multiplyMatricesretVect(partMatrixA, matrixB, rows, colsA, rowsB, colsB, offset, elements);   /*Buffer used to store computed values to send to MASTER*/ 
+    long int *recvGatBuf = malloc(sizeof(long int) * elements * tasksNum);
     int recvCounts[tasksNum];
     int displs[tasksNum];
     displs[0] = 0;
@@ -133,13 +133,13 @@ int main(int argc, char* argv[]) {
         displs[taskRank+1] = displs[taskRank] + recvCounts[taskRank]; 
     }
 
-    MPI_Gatherv(sendBuf, elements, MPI_INT, recvGatBuf, recvCounts, displs, MPI_INT, MASTER, MPI_COMM_WORLD);
+    MPI_Gatherv(sendBuf, elements, MPI_LONG, recvGatBuf, recvCounts, displs, MPI_LONG, MASTER, MPI_COMM_WORLD);
 
     if(taskId == MASTER){
         FILE *outFile = fopen("outfile", "w+");
         fprintf(outFile, "%d %d\n", rowsA, colsB);
         for(int i = 0; i < rowsA * colsB; i++){
-            fprintf(outFile, "%d ", recvGatBuf[i]);
+            fprintf(outFile, "%ld ", recvGatBuf[i]);
             if((i + 1) % colsB == 0)
                 fprintf(outFile, "\n");
         }
